ETL-Pipeline-main/requirements.txt
functions-framework>=3.6.0
SQLAlchemy>=2.0.0,<3.0.0
PyMySQL>=1.1.0
google-cloud-storage>=2.12.0
cloud-sql-python-connector>=1.4.0
numpy>=1.21.0,<2.0.0
pandas>=1.5.0,<2.0.0
gcsfs>=2023.6.0
fsspec>=2023.6.0
facebook_business>=20.0.0
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.66.0
google-api-python-client>=2.98.0
mysql-connector-python>=8.3.0
google-ads>=21.3.0
google-auth-oauthlib>=1.2.0
prefect>=2.13.0
